requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
web3==5.25.0
python-dotenv==0.17.1
